fix: resolve critical bugs across 12 crates (v1.3.0)
Protocol fixes:
- MySQL: fix binary protocol integer type sizes, USE command case sensitivity
- Redis: fix inline command first char parsing, add negative TTL validation
- ClickHouse: fix delete_where column length panic, HTTP status codes, DROP TABLE IF EXISTS
- MongoDB: fix $group field reference match arm order, $match pipeline, $exists null handling

Storage/DataFusion fixes:
- fe-storage: add write locks for drop_table/create_table, fix filter pushdown error handling
- fe-datafusion: fix UInt64→Int64 truncation, date_add i64→i32 truncation, unix_timestamp volatility
- fe-sql-parser: fix COMMIT; parsing, CATALOG case sensitivity, INSERT SET case sensitivity

Security fixes:
- MySQL auth: add random JWT secret fallback, require JWT_SECRET env var in production
- fe-catalog: fix test_rocks_backend backend name mismatch

Tests added:
- 8 new test cases for fe-sql-parser bug fixes
- All 850+ unit tests pass

+++ crates/mysql-protocol/src/packet.rs
@@ -1164,9 +1164,9 @@ fn encode_binary_value(pb: &mut PacketBuilder, val: &BinaryValue) {
 pub fn text_to_binary(text: Option<&str>, col_type: u8) -> Option<BinaryValue> {
     let text = text?;
     match col_type {
-        column_type::TINY => text.parse::<i64>().ok().map(|n| BinaryValue::Int64(n)),
-        column_type::SHORT => text.parse::<i64>().ok().map(|n| BinaryValue::Int64(n)),
-        column_type::LONG => text.parse::<i64>().ok().map(|n| BinaryValue::Int64(n)),
+        column_type::TINY => text.parse::<i8>().ok().map(|n| BinaryValue::Int8(n)),
+        column_type::SHORT => text.parse::<i16>().ok().map(|n| BinaryValue::Int16(n)),
+        column_type::LONG => text.parse::<i32>().ok().map(|n| BinaryValue::Int32(n)),
         column_type::LONGLONG => text.parse::<i64>().ok().map(|n| BinaryValue::Int64(n)),
         column_type::FLOAT => text.parse::<f64>().ok().map(|f| BinaryValue::Double(f)),
         column_type::DOUBLE => text.parse::<f64>().ok().map(|f| BinaryValue::Double(f)),
